backend/migration_note.py
import os
from dotenv import load_dotenv
from supabase import create_client

load_dotenv()

url = os.environ.get("SUPABASE_URL")
key = os.environ.get("SUPABASE_KEY")
supabase = create_client(url, key)

# SQL to add column if not exists (Postgres specific)
sql = """
ALTER TABLE movies 
ADD COLUMN IF NOT EXISTS ott text DEFAULT 'Netflix';
"""

try:
    # We can't run DDL via client easily for all versions, but let's try via rpc or just raw sql if supported, 
    # but supabase-py client usually runs queries on tables. 
    # Since we don't have direct SQL access, we will use the user's dashboard instructions normally.
    # HOWEVER, for this agent, we can try to re-run schema.sql or just assume the user will run it.
    # Wait, the previous seed used plain text insert.
    
    # Actually, we can use the backend execution if we had a raw SQL endpoint, but we don't.
    # I will have to ASK the user to run SQL or try to run it via the 'postgres' connection string if I had it.
    # But I only have API keys. 
    
    # WORKAROUND: I cannot alter table via Supabase JS/Python Client easily without specific privileges or RPC.
    # BUT, I can ask the user to paste SQL.
    # OR better: I will try to update the schema through the known 'schema.sql' and assume user runs it?
    # No, that's bad UX.
    
    # Let's try to infer if we can just "add" it to keys in insert and see if Supabase auto-adds? No, it's SQL.
    pass
except Exception as e:
    print(e)
